Final hoisting and file cleanup
