+++ packages/shared-ui/components/empty-block.tsx
@@ -2,12 +2,14 @@ import { FilePlus2, LucideProps } from "lucide-react";
 
 import { cn } from "@shared/lib/utils";
 import { Button } from "@shared/shadcn/ui/button";
+import { Skeleton } from "@shared/shadcn/ui/skeleton";
 
 interface Props {
     title: string;
     description?: string;
     buttonLabel?: string;
     onButtonClick?: () => void;
+    isLoading?: boolean;
     Icon?: React.ForwardRefExoticComponent<
         Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>
     >;
@@ -19,9 +21,14 @@ export const EmptyBlock = ({
     title,
     buttonLabel,
     onButtonClick,
+    isLoading = false,
     Icon = FilePlus2,
     iconClass = "",
 }: Props) => {
+    if (isLoading) {
+        return <Skeleton className="h-[450px] w-full rounded-md" />;
+    }
+
     return (
         <div className="flex h-[450px] shrink-0 items-center justify-center rounded-md border border-dashed">
             <div className="mx-auto flex max-w-[420px] flex-col items-center justify-center text-center">
