+++ packages/user/FractalCore/ui/src/components/current-evaluation-card.tsx
@@ -2,7 +2,8 @@ import { useWatchAttest } from "@/hooks/fractals/use-watch-attest";
 import { useWatchClose } from "@/hooks/fractals/use-watch-close";
 import { useWatchStart } from "@/hooks/fractals/use-watch-start";
 import { useEvaluationStatus } from "@/hooks/use-evaluation-status";
-import { useNowUnix } from "@/hooks/use-now-unix";
+
+import { useNowUnix } from "@shared/hooks/use-now-unix";
 
 import { Deliberation } from "./evaluations/deliberation";
 import { Failed } from "./evaluations/failed";
